requirements.txt
tweepy
python-decouple
arxiv
rmapy
notion_df
pandas